requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
et_xmlfile==2.0.0
execnet==2.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-dependency==0.6.0
pytest-fixture-marker==1.0.0
pytest-fixture-order==0.1.4
pytest-fixtures==0.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-selenium==4.1.0
pytest-variables==3.1.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
